chore: webhook to typescript
